backend/requirements.txt
google-generativeai
python-dotenv
paramiko
reportlab
Flask
pyyaml
Flask-CORS